requirements.txt
DateTime==4.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.5
django-braces==1.8.1
django-filter==0.13.0
django-oauth-toolkit==0.10.0
django-orm==2.0.5
django-toolbelt==0.0.1
djangorestframework==3.3.3
google-api-python-client==1.5.1
gunicorn==19.4.5
httplib2==0.9.2
Jinja==1.2
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==0.23
oauth2==1.9.0.post1
oauth2client==2.1.0
oauthlib==1.0.3
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycurl==7.43.0
PyJWT==1.4.0
python-gflags==3.0.5
python-openid==2.2.5
python-social-auth==0.2.19
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.1
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
static3==0.6.1
uritemplate==0.6
whitenoise==3.0
zope.interface==4.1.3
